refactor(nuqs-jotai): simplify binding initialization and organization
