+++ packages/nuqs-jotai/README.md
@@ -72,7 +72,7 @@ forwarded to nuqs; the atom snapshot reflects nuqs's next React render. Do not
 expect a `get` immediately following `set` in the same command to read a new
 snapshot. Functional updates are passed directly to nuqs and follow its batching
 semantics. A write promise is nuqs's URL-commit promise, not a React-render or
-server-render completion signal.
+server-render completion signal. Handle write rejections at the command owner.
 
 The bridge holds mount/reveal commands until nuqs's subscriptions are attached,
 then forwards them unchanged in a microtask after the current effect pass. This
@@ -130,7 +130,7 @@ example, `parseAsNativeArrayOf(parseAsString)` cannot distinguish an empty array
 from an empty string in the committed URL; nuqs may normalize `[]` to `['']` after
 committing `?key=`. The bridge does not override that behavior.
 
-## Testing and migration
+## Testing
 
 `QueryTestingAdapter` from `nuqs-jotai/testing` wraps the real NuqsTestingAdapter
 and accepts its props plus `atoms`. Use `hasMemory` when checking committed URL
@@ -141,18 +141,6 @@ state. Wrap React-driven writes in Testing Library `act`.
   <Search />
 </QueryTestingAdapter>
 ```
-
-From the previous unmerged API:
-
-- Replace `QueryStateProvider adapter={...}` with `atoms={[...]}` beneath the
-  application's native NuqsAdapter. Remove `createBrowserQueryAdapter`, custom
-  refresh callbacks and `notifyUrlChange` wiring.
-- Move provider options to NuqsAdapter's `defaultOptions`.
-- Replace URL `useAtomValue` / `useAtom` reads with `useQueryAtomValue` /
-  `useQueryAtom`; keep `useSetAtom` and derived/write atom definitions.
-- Replace custom memory adapters with QueryTestingAdapter or NuqsTestingAdapter.
-- Handle write rejections at the command owner. The provider-local
-  `queryStateErrorAtom` and custom rollback/cancellation policy were removed.
 
 Run `pnpm --filter nuqs-jotai test run` and
 `pnpm --filter nuqs-jotai type-check`. Tests cover native nuqs interoperability,
